Implement getting profile if not on team
